TST/REF: collect indexing tests by method (#37638)
