+++ pandas/tests/frame/indexing/test_sparse.py
@@ -1,12 +1,6 @@
-import numpy as np
-import pytest
-
-import pandas.util._test_decorators as td
-
 import pandas as pd
 import pandas._testing as tm
 from pandas.arrays import SparseArray
-from pandas.core.arrays.sparse import SparseDtype
 
 
 class TestSparseDataFrameIndexing:
@@ -23,34 +17,3 @@ class TestSparseDataFrameIndexing:
 
         result = df.loc[:, "A"]
         tm.assert_series_equal(result, expected)
-
-    @pytest.mark.parametrize("spmatrix_t", ["coo_matrix", "csc_matrix", "csr_matrix"])
-    @pytest.mark.parametrize("dtype", [np.int64, np.float64, complex])
-    @td.skip_if_no_scipy
-    def test_loc_getitem_from_spmatrix(self, spmatrix_t, dtype):
-        import scipy.sparse
-
-        spmatrix_t = getattr(scipy.sparse, spmatrix_t)
-
-        # The bug is triggered by a sparse matrix with purely sparse columns.  So the
-        # recipe below generates a rectangular matrix of dimension (5, 7) where all the
-        # diagonal cells are ones, meaning the last two columns are purely sparse.
-        rows, cols = 5, 7
-        spmatrix = spmatrix_t(np.eye(rows, cols, dtype=dtype), dtype=dtype)
-        df = pd.DataFrame.sparse.from_spmatrix(spmatrix)
-
-        # regression test for #34526
-        itr_idx = range(2, rows)
-        result = df.loc[itr_idx].values
-        expected = spmatrix.toarray()[itr_idx]
-        tm.assert_numpy_array_equal(result, expected)
-
-        # regression test for #34540
-        result = df.loc[itr_idx].dtypes.values
-        expected = np.full(cols, SparseDtype(dtype, fill_value=0))
-        tm.assert_numpy_array_equal(result, expected)
-
-    def test_all_sparse(self):
-        df = pd.DataFrame({"A": pd.array([0, 0], dtype=pd.SparseDtype("int64"))})
-        result = df.loc[[0, 1]]
-        tm.assert_frame_equal(result, df)
