+++ pandas/tests/indexing/test_datetime.py
@@ -1,6 +1,3 @@
-from datetime import date, datetime, timedelta
-
-from dateutil import tz
 import numpy as np
 import pytest
 
@@ -206,26 +203,6 @@ class TestDatetimeIndex:
         df.loc[mask, "C"] = df.loc[mask].index
         tm.assert_frame_equal(df, expected)
 
-    def test_loc_setitem_datetime(self):
-
-        # GH 9516
-        dt1 = Timestamp("20130101 09:00:00")
-        dt2 = Timestamp("20130101 10:00:00")
-
-        for conv in [
-            lambda x: x,
-            lambda x: x.to_datetime64(),
-            lambda x: x.to_pydatetime(),
-            lambda x: np.datetime64(x),
-        ]:
-
-            df = DataFrame()
-            df.loc[conv(dt1), "one"] = 100
-            df.loc[conv(dt2), "one"] = 200
-
-            expected = DataFrame({"one": [100.0, 200.0]}, index=[dt1, dt2])
-            tm.assert_frame_equal(df, expected)
-
     def test_series_partial_set_datetime(self):
         # GH 11497
 
@@ -245,7 +222,8 @@ class TestDatetimeIndex:
         exp = Series(
             [0.2, 0.2, 0.1], index=pd.DatetimeIndex(keys, name="idx"), name="s"
         )
-        tm.assert_series_equal(ser.loc[keys], exp, check_index_type=True)
+        result = ser.loc[keys]
+        tm.assert_series_equal(result, exp, check_index_type=True)
 
         keys = [
             Timestamp("2011-01-03"),
@@ -273,7 +251,8 @@ class TestDatetimeIndex:
             pd.Period("2011-01-01", freq="D"),
         ]
         exp = Series([0.2, 0.2, 0.1], index=pd.PeriodIndex(keys, name="idx"), name="s")
-        tm.assert_series_equal(ser.loc[keys], exp, check_index_type=True)
+        result = ser.loc[keys]
+        tm.assert_series_equal(result, exp, check_index_type=True)
 
         keys = [
             pd.Period("2011-01-03", freq="D"),
@@ -297,33 +276,6 @@ class TestDatetimeIndex:
         expected = DataFrame(-1, index=index, columns=["a"])
         tm.assert_frame_equal(result, expected)
 
-    def test_loc_getitem_across_dst(self):
-        # GH 21846
-        idx = pd.date_range(
-            "2017-10-29 01:30:00", tz="Europe/Berlin", periods=5, freq="30 min"
-        )
-        series2 = Series([0, 1, 2, 3, 4], index=idx)
-
-        t_1 = Timestamp("2017-10-29 02:30:00+02:00", tz="Europe/Berlin", freq="30min")
-        t_2 = Timestamp("2017-10-29 02:00:00+01:00", tz="Europe/Berlin", freq="30min")
-        result = series2.loc[t_1:t_2]
-        expected = Series([2, 3], index=idx[2:4])
-        tm.assert_series_equal(result, expected)
-
-        result = series2[t_1]
-        expected = 2
-        assert result == expected
-
-    def test_loc_incremental_setitem_with_dst(self):
-        # GH 20724
-        base = datetime(2015, 11, 1, tzinfo=tz.gettz("US/Pacific"))
-        idxs = [base + timedelta(seconds=i * 900) for i in range(16)]
-        result = Series([0], index=[idxs[0]])
-        for ts in idxs:
-            result.loc[ts] = 1
-        expected = Series(1, index=idxs)
-        tm.assert_series_equal(result, expected)
-
     def test_loc_setitem_with_existing_dst(self):
         # GH 18308
         start = Timestamp("2017-10-29 00:00:00+0200", tz="Europe/Madrid")
@@ -339,39 +291,3 @@ class TestDatetimeIndex:
             dtype=object,
         )
         tm.assert_frame_equal(result, expected)
-
-    def test_loc_str_slicing(self):
-        ix = pd.period_range(start="2017-01-01", end="2018-01-01", freq="M")
-        ser = ix.to_series()
-        result = ser.loc[:"2017-12"]
-        expected = ser.iloc[:-1]
-
-        tm.assert_series_equal(result, expected)
-
-    def test_loc_label_slicing(self):
-        ix = pd.period_range(start="2017-01-01", end="2018-01-01", freq="M")
-        ser = ix.to_series()
-        result = ser.loc[: ix[-2]]
-        expected = ser.iloc[:-1]
-
-        tm.assert_series_equal(result, expected)
-
-    @pytest.mark.parametrize(
-        "slice_, positions",
-        [
-            [slice(date(2018, 1, 1), None), [0, 1, 2]],
-            [slice(date(2019, 1, 2), None), [2]],
-            [slice(date(2020, 1, 1), None), []],
-            [slice(None, date(2020, 1, 1)), [0, 1, 2]],
-            [slice(None, date(2019, 1, 1)), [0]],
-        ],
-    )
-    def test_getitem_slice_date(self, slice_, positions):
-        # https://github.com/pandas-dev/pandas/issues/31501
-        s = Series(
-            [0, 1, 2],
-            pd.DatetimeIndex(["2019-01-01", "2019-01-01T06:00:00", "2019-01-02"]),
-        )
-        result = s[slice_]
-        expected = s.take(positions)
-        tm.assert_series_equal(result, expected)
