+++ pandas/tests/indexing/test_partial.py
@@ -351,31 +351,6 @@ class TestPartialSetting:
         tm.assert_index_equal(df.index, Index(orig.index.tolist() + ["a"]))
         assert df.index.dtype == "object"
 
-    def test_partial_set_empty_series(self):
-
-        # GH5226
-
-        # partially set with an empty object series
-        s = Series(dtype=object)
-        s.loc[1] = 1
-        tm.assert_series_equal(s, Series([1], index=[1]))
-        s.loc[3] = 3
-        tm.assert_series_equal(s, Series([1, 3], index=[1, 3]))
-
-        s = Series(dtype=object)
-        s.loc[1] = 1.0
-        tm.assert_series_equal(s, Series([1.0], index=[1]))
-        s.loc[3] = 3.0
-        tm.assert_series_equal(s, Series([1.0, 3.0], index=[1, 3]))
-
-        s = Series(dtype=object)
-        s.loc["foo"] = 1
-        tm.assert_series_equal(s, Series([1], index=["foo"]))
-        s.loc["bar"] = 3
-        tm.assert_series_equal(s, Series([1, 3], index=["foo", "bar"]))
-        s.loc[3] = 4
-        tm.assert_series_equal(s, Series([1, 3, 4], index=["foo", "bar", 3]))
-
     def test_partial_set_empty_frame(self):
 
         # partially set with an empty object
@@ -504,10 +479,12 @@ class TestPartialSetting:
         # GH 5756
         # setting with empty Series
         df = DataFrame(Series(dtype=object))
-        tm.assert_frame_equal(df, DataFrame({0: Series(dtype=object)}))
+        expected = DataFrame({0: Series(dtype=object)})
+        tm.assert_frame_equal(df, expected)
 
         df = DataFrame(Series(name="foo", dtype=object))
-        tm.assert_frame_equal(df, DataFrame({"foo": Series(dtype=object)}))
+        expected = DataFrame({"foo": Series(dtype=object)})
+        tm.assert_frame_equal(df, expected)
 
     def test_partial_set_empty_frame_empty_copy_assignment(self):
         # GH 5932
@@ -565,19 +542,17 @@ class TestPartialSetting:
         ],
     )
     def test_loc_with_list_of_strings_representing_datetimes(
-        self, idx, labels, expected_idx
+        self, idx, labels, expected_idx, frame_or_series
     ):
         # GH 11278
-        s = Series(range(20), index=idx)
-        df = DataFrame(range(20), index=idx)
+        obj = frame_or_series(range(20), index=idx)
 
         expected_value = [3, 7, 11]
-        expected_s = Series(expected_value, expected_idx)
-        expected_df = DataFrame(expected_value, expected_idx)
+        expected = frame_or_series(expected_value, expected_idx)
 
-        tm.assert_series_equal(expected_s, s.loc[labels])
-        tm.assert_series_equal(expected_s, s[labels])
-        tm.assert_frame_equal(expected_df, df.loc[labels])
+        tm.assert_equal(expected, obj.loc[labels])
+        if frame_or_series is Series:
+            tm.assert_series_equal(expected, obj[labels])
 
     @pytest.mark.parametrize(
         "idx,labels",
@@ -651,16 +626,6 @@ class TestPartialSetting:
         with pytest.raises(KeyError, match=msg):
             df.loc[labels]
 
-    def test_indexing_timeseries_regression(self):
-        # Issue 34860
-        arr = date_range("1/1/2008", "1/1/2009")
-        result = arr.to_series()["2008"]
-
-        rng = date_range(start="2008-01-01", end="2008-12-31")
-        expected = Series(rng, index=rng)
-
-        tm.assert_series_equal(result, expected)
-
     def test_index_name_empty(self):
         # GH 31368
         df = DataFrame({}, index=pd.RangeIndex(0, name="df_index"))
@@ -688,17 +653,4 @@ class TestPartialSetting:
         df = DataFrame(range(len(index)), index=index)
         expected = DataFrame(range(len(index[:3])), index=index[:3])
         result = df["2012-01-01":"2012-01-04"]
-        tm.assert_frame_equal(result, expected)
-
-    def test_slice_datetime_index(self):
-        # GH35509
-        df = DataFrame(
-            {"col1": ["a", "b", "c"], "col2": [1, 2, 3]},
-            index=pd.to_datetime(["2020-08-01", "2020-07-02", "2020-08-05"]),
-        )
-        expected = DataFrame(
-            {"col1": ["a", "c"], "col2": [1, 3]},
-            index=pd.to_datetime(["2020-08-01", "2020-08-05"]),
-        )
-        result = df.loc["2020-08"]
         tm.assert_frame_equal(result, expected)
