+++ pandas/tests/indexing/test_timedelta.py
@@ -104,19 +104,3 @@ class TestTimedeltaIndexing:
 
         assert expected == result
         tm.assert_frame_equal(df, df_copy)
-
-    def test_loc_str_slicing(self):
-        ix = pd.timedelta_range(start="1 day", end="2 days", freq="1H")
-        ser = ix.to_series()
-        result = ser.loc[:"1 days"]
-        expected = ser.iloc[:-1]
-
-        tm.assert_series_equal(result, expected)
-
-    def test_loc_slicing(self):
-        ix = pd.timedelta_range(start="1 day", end="2 days", freq="1H")
-        ser = ix.to_series()
-        result = ser.loc[: ix[-2]]
-        expected = ser.iloc[:-1]
-
-        tm.assert_series_equal(result, expected)
