+++ pandas/tests/series/indexing/test_numeric.py
@@ -71,17 +71,6 @@ def test_getitem_setitem_slice_integers():
     assert not (s[4:] == 0).any()
 
 
-def test_setitem_float_labels():
-    # note labels are floats
-    s = Series(["a", "b", "c"], index=[0, 0.5, 1])
-    tmp = s.copy()
-
-    s.loc[1] = "zoo"
-    tmp.iloc[2] = "zoo"
-
-    tm.assert_series_equal(s, tmp)
-
-
 def test_slice_float_get_set(datetime_series):
     msg = (
         "cannot do slice indexing on DatetimeIndex with these indexers "
